refactor: Rename Files and Adjust Folder Structure
to improve code clarity

Files Renamed
- Task -> TaskCard
- Tasks -> TaskList
- SingleTask -> TaskDetails
- Dashboard -> TasksDashboard
- Offers -> OffersList

Folders Renamed and Files Moved
- ViewTask folder renamed to TaskList
- TaskDetails, OffersLists & Offer moved into TaskDetails folder
