+++ components/tasks/TaskDetails/index.tsx
@@ -1,0 +1,3 @@
+export * from "./TaskDetails";
+export * from "./OffersList";
+export * from "./Offer";
