+++ components/tasks/TaskList/index.tsx
@@ -1,0 +1,3 @@
+export * from "./TasksList"
+export * from "./FilterTasks"
+export * from "./TaskCard"
