+++ pages/mytasks/index.tsx
@@ -1,8 +1,8 @@
 import type { NextPage } from "next";
-import Dashboard from "../../components/tasks/Dashboard";
+import TasksDashboard from "../../components/tasks/TasksDashboard";
 
 const MyTasksPage: NextPage = () => {
-  return <Dashboard myTasks={true} />;
+  return <TasksDashboard myTasks={true} />;
 };
 
 export default MyTasksPage;
